Improve tournament and player name detection and change player value to a float.

+++ db/migrate/20161116110416_create_players.rb
@@ -5,7 +5,7 @@ class CreatePlayers < ActiveRecord::Migration[5.0]
       t.string :role
       t.references :team
       t.string :country
-      t.integer :value,   null: false, default: 100
+      t.float :value,     null: false, default: 100.0
       t.string :slug,     null: false
 
       t.timestamps
